need mock dhcpserver instance
