+++ tests/host/Makefile
@@ -143,6 +143,7 @@ MOCK_CPP_FILES_EMU := $(MOCK_CPP_FILES_COMMON) \
 		ArduinoMainUdp.cpp \
 		ArduinoMainSpiffs.cpp \
 		ArduinoMainLittlefs.cpp \
+		DhcpServer.cpp \
 		user_interface.cpp \
 	)
